move the homepage route component into its own folder and create a styel component, which structures the trending sectino on mobile devices
